+++ R/eval_replacement_importance.R
@@ -433,10 +433,8 @@ methods::setMethod("eval_replacement_importance",
     }
     # return result
     out <- tibble::as_tibble(as.data.frame(m_total))
-    sf::st_as_sf(
-      out, geometry = sf::st_geometry(x$data$cost),
-      crs = sf::st_crs(x$data$cost)
-    )
+    out$geometry <- sf::st_geometry(x$data$cost)
+    sf::st_sf(out, crs = sf::st_crs(x$data$cost))
   }
 )